Small optmisation to day 03 solution
The second set don't need to contain all items in the 2nd rucksack,
only the ones already present in the 1st one. This way bcBoth (cTwo before)
before is now the effective set union of 1st and 2nd rucksacks and bc1 (cOne
before) does not need to be used in the final iteration in the 3rd rucksack.