Day 19: Return/Save only the best result on BT
We were saving and returning the whole state during the backtrack (which is
useful to debug it), but in order to compute the optimal solution we only need
to keep track of the geodes in that state. Using this reduces the the space required
for the values of the map by a factor of ~9.